Greedy.jl

Greedy

Greedy.Partition — Type.

Disjoint-Set structure für Kruskal

source

Greedy.generate_matroid_question — Function.

Generiert eine Frage, in der eine Minimalbasis bestimmt werden soll

source

Greedy.generate_spantree_question — Function.

Generiert eine Frage, in der ein minimaler Spannbaum bestimmt werden soll

source

Greedy.kruskal — Method.

Algorithmus von Kruskal

source

Greedy.random_spantree — Method.

Generiert einen zufälligen Spannbaum – Kruskal mit zufälligen Gewichten

source

Greedy.uniqueify_matroid — Function.

Algorithmus 2.2: Generiert eine Kostenfunktion für das Matroid, sodass $B$ die eindeutige Minimalbasis ist

source

Greedy.uniqueify_spantree — Method.

Algorithmus 2.2: Generiert eine Kostenfunktion für den Graphen, sodass $T$ der eindeutige Minimale Spannbaum ist.

source

Greedy.max_edge_in_path — Function.

Hilfsfunktion für random_spantree, im Prinzip eine DFS. Findet das teuerste Gewicht auf dem (vorausgesetzt eindeutigen) $s$-$t$-Weg.

source