ADMStructures.jl

ADMStructures

Hier finden sich allgemeine Strukturen der ADM, wie Matroide und (Di-)Graphen.

Matroide

ADMStructures.Matroid — Type.

Ein Matroid in abstrakter Repräsentation. Wird initialisiert durch das Basensystem:

Variablen:

E: Grundmenge des Matroids

I: Unabhängige Mengen des Matroids

julia> m = Matroid(bases=[(1, 2, 3), (2, 3, 4)])
ADMStructures.Matroid(Any[1, 2, 3, 4], Any[Int64[], [1], [2], [3], [4], [1, 2], [1, 3], [2, 3], [2, 4], [3, 4], [1, 2, 3], [2, 3, 4]])

source

ADMStructures.is_matroid — Function.

Prüft, ob m tatsächlich ein Matroid ist.

julia> is_matroid(m)
true

source

ADMStructures.matroid_graph — Function.

Gibt eine hierarchische Repräsentation des Matroids als Tikz-Picture zurück.

source

ADMStructures.rank — Function.

Gibt den Rang eines Matroids zurück

julia> ADMStructures.rank(m)
3

source

ADMStructures.bases — Function.

Gibt die Basen eines Matroids zurück

julia> bases(m)
2-element Array{Array{Int64,1},1}:
 [1, 2, 3]
 [2, 3, 4]

source

ADMStructures.circles — Function.

Gibt die Kreise eines Matroids zurück

julia> circles(m)
1-element Array{Any,1}:
 [1, 4]

source

Graphen

ADMStructures.Graph — Type.

Enthält Information über einen Graphen, gerichtet oder ungerichtet.

Variablen:

V: Die Knoten des Graphen

E: Die Kanten des Graphen

positions: Die Position der einzelnen Grahen – Wird per Federnetzwerk bestimmt, falls nicht angegeben.

labels: LaTeX-Labels für die einzelnen Knoten

directed: Boolean, der angibt, ob der Graph gerichtet ist

source

ADMStructures.graph — Function.

Nimmt einen Graphen und eine Kostenfunktion entgegen und wandelt den Graphen in ein TikzPicture um.

source

ADMStructures.graph_svg — Function.

Funktionsweise wie graph, gibt SVG-Rohdaten zurück.

source

ADMStructures.graph_moodle — Function.

Funktionsweise wie graph, gibt ein MoodleFile statt ein TikzPicture zurück.

source

ADMStructures.spring_positions — Function.

Findet das bestmögliche rechteckige Federnetzwerk, indem folgendes fixiert ist:

der erste Knoten auf $(0, 0)$

der letzte Knoten auf $(w, 0)$

ein Knoten auf $(\cdot, h/2)$

ein Knoten auf $(\cdot, -h/2)$

Alle Möglichkeiten für die letzten beiden Knoten werden in Betracht gezogen und die beste Kombination gewählt.

source

ADMStructures.spring_positions! — Function.

Inplace-Version von spring_positions

source

ADMStructures.build_mesh_graph — Function.

Generiert einen rechteckigen Gittergraphen.

source

ADMStructures.build_wheel_graph — Function.

Generiert einen Radgraphen mit n Speichen.

source

ADMStructures.greedy_labelling! — Function.

Gibt den Knoten des Graphen Label, wie es für Minimale Spannbäume üblich ist. Alle Knoten erhalten ein Label $v_i$.

source

ADMStructures.sp_labelling! — Function.

Gibt den Knoten des Graphen Label, wie es für Kürzeste Wege / Max Flow üblich ist. Der erste Knoten wird mit $s$ gelabelt, der letzte mit $t$. Alle anderen Knoten erhalten ein Label $v_i$.

source

Allgemeine Funktionen

ADMStructures.powerset — Function.

Gibt die Potenzmenge zurück.

julia> powerset([1, 2, 3])
8-element Array{Any,1}:
 Int64[]  
 [1]      
 [2]      
 [3]      
 [1, 2]   
 [1, 3]   
 [2, 3]   
 [1, 2, 3]

source

ADMStructures.argmin — Function.

Gibt den Index eines niedrigsten Wertes zurück. Bestimmt wahlweise auch, ob dieser Index eindeutig ist.

julia> argmin([1, 2, 3], return_uniq=true)
(1, true)

source

ADMStructures.random_order — Function.

Permutiert eine Liste zufällig. julia> srand(42); random_order([1, 2, 3, 4]) 4-element Array{Int64,1}: 4 3 1 2

source

ADMStructures.repr_html — Function.

HTML-Repräsentation eines ADM-Objekts

source

ADMStructures.repr_tex — Function.

LaTeX-Repräsentation eines ADM-Objekts

source