Knapsack.jl

Knapsack

Knapsack.alpha — Method.

Die $\alpha$-Funktion des Knapsack-Algorithmus. Memoisiert, um dynamische Optimierung zu implementieren. Gibt nicht nur den Nutzen, sondern auch die bisher verwendeten Objekte und den in der Arbeit beschriebenen $\gamma$-Wert zurück.

source

Knapsack.generate_knapsack_question — Method.

Generiert eine Frage, in der ein optimaler Knapsack bestimmt werden soll

source

Knapsack.set_string — Function.

Wandelt ein Array in das Format {1, 2, 3, ...} um

julia> set_string([1, 2, 3, 4])
"{1, 2, 3, 4}"

source

Knapsack.unzip — Method.

Entzippt eine Liste von Tupeln. Umkehrfunktion von zip. Implementierung angepasst von diesem Github Issue.

source