MaxFlow.jl

MaxFlow

MaxFlow.generate_maxflow_question — Method.

Generiert eine Frage, in der ein Max Flow bestimmt werden soll

source

MaxFlow.has_circle — Function.

Überprüft, ob ein gegebener Graph einen Kreis hat. In gerichteten Graphen prüft er den Graphen auf gerichtete Kreise, analog für ungerichtete Graphen.

source

MaxFlow.uniqueify_network — Method.

Implementierung von Algorithmus 5.2 aus der Arbeit

source